refactor: move business logic out of imageGen/pipeline routes into services
From the 2026-06-09 audit. The 214-line imageGen /generate handler
delegates to a new services/imageGen/prepareParams.js; the clean /
de-watermark / light-regen variant endpoints share a new
services/imageGen/variants.js (persist-variant tail deduplicated); the
SSE writeHead/send/safeEnd boilerplate in imageGen/videoGen collapses
onto a shared openSseStream helper in sseDownload.js. llmSchema moves
from a sibling-route import into lib/validation.js, episode canon
extraction moves from the pipeline route into arcPlanner, and the
/templates page registers in the nav manifest so it is reachable from
the command palette and voice navigation. imageGen tests pin the avatar
response shape and add a regenerate happy-path contract.
